refactor(infrastructure): PR-7 commit 1 — 5 leaf 模块搬 infrastructure/
paths / event_bus / log_tail / argparse_bridge / llm_presets 5 文件搬到
studio/infrastructure/，老路径留 3 行 sys.modules 别名 shim 保兼容。

  studio/infrastructure/__init__.py   13   说明文档（含 PR-7 c2/c3 计划）
  studio/infrastructure/paths.py       88   路径常量 + safe_join / validate_path_component
  studio/infrastructure/event_bus.py   90   进程内 SSE 总线
  studio/infrastructure/log_tail.py   240   per-task 日志增量读 + monitor state 轮询
  studio/infrastructure/argparse_bridge.py 186 pydantic → argparse
  studio/infrastructure/llm_presets.py  41   builtin LLM caption preset 加载

  studio/{paths,event_bus,log_tail,argparse_bridge,llm_presets}.py  各 ~10 行 shim
    每个 shim 同 PR-3 services/ 模式：
      import sys as _sys
      from .infrastructure import X as _real
      _sys.modules[__name__] = _real
    让旧路径 `from studio.paths import X` / monkeypatch.setattr(paths, X) 透明。

**关键修复**：搬到 infrastructure/ 后两个 file-relative 路径要补一层 .parent：
  - paths.py: `REPO_ROOT = Path(__file__).resolve().parent.parent.parent`
    （以前 .parent.parent；现在 __file__ → infrastructure/ → studio/ → repo root）
  - llm_presets.py: `PRESETS_DIR = .parent.parent / "llm_presets"`
    （json 数据仍在 studio/llm_presets/，不搬）

回归：PR-1 安全网 152 + studio_server 44 + supervisor 全套 = **265/265 全绿**。
route snapshot 不变（160 routes 含 SPA mount）。
